# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Store API",
        "description": "A store listing API based on Swagger and built using Java"
    },
    "host": "localhost",
    "schemes": [
        "http",
        "https"
    ],
    "basePath": "/api",
    "paths": {
...
        "/Albums/{id}": {
            "get": {
                "tags": [
                    "Album"
                ],
                "operationId": "albums_getById",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Album"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/Albums?artistId={id}": {
            "get": {
                "tags": [
                    "Album",
                    "Artist"
                ],
                "operationId": "albums_getByArtistId",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Album"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "Album": {
            "type": "object",
            "properties": {
                "ID": {
                    "format": "int32",
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Artist": {
                    "format": "int32",
                    "type": "integer"
                },
                "Url": {
                    "type": "string"
                },
                "Image": {
                    "type": "string"
                },
                "Listeners": {
                    "type": "string"
                },
                "Playcount": {
                    "type": "string"
                }
            }
        },
...
    }
}